+++ Sources/ZincFramework/Commands/TestCommand.swift
@@ -1,5 +1,7 @@
 // Copyright © 2019 SpotHero. All rights reserved.
 
+import CommandHero
+
 import Foundation
 
 class TestCommand: Command {
